pyproject.toml
[project]
name = "BigData_project"
version = "0.1.0"
description = "NYC Taxi Trip Streaming Pipeline: Kafka → Spark → Cassandra → Grafana"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cassandra-driver>=3.29.3",
    "confluent-kafka[avro]>=2.12.0",
    "fastapi>=0.128.0",
    "polars>=1.34.0",
    "pyspark==3.5.7",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
]
